Added PowerTest
This test includes a script (PowerTest.py), more configuration variables
in Config.py, a bit of documentation on how to use it, and a default
Table F for input.  Note that it is not completely done because it does
not set the inital conditions for the frequency controller, due to the
fact that we could not implement it in the lab yet.