Started I2C connection with load control arduino
This is just the current progress (not working yet) on the connection
between the Raspberry Pi that run the test scripts and the Arduino that
sets loads via I2C.  The Arduino code is a test system and the changes
will later be merged into the actual load control when it is useful. I
also modified the interface a bit - by adding a null character at the
end of command transition, but that may be unecessary and I will work to
change it back if possible.